refactor: move CallbackTable to separate file
