+++ pkg/sentry/kernel/callbacks.go
@@ -8,7 +8,6 @@ import (
 	"gvisor.dev/gvisor/pkg/sentry/kernel/callbacks"
 	"strconv"
 	"strings"
-	"sync"
 )
 
 type ContextAddable interface {
@@ -109,127 +108,6 @@ type CallbackAfter interface {
 
 	// Info about this callback
 	Info() callbacks.JsCallbackInfo
-}
-
-// CallbackTable is a storage of functions which can be called before and/ or after syscall execution
-// TODO incapsulate the mutex (exposing mutex - straight way to deadlock or other memes)
-type CallbackTable struct {
-	// callbackBefore is a map of:
-	//	key - sysno (uintptr)
-	//	val - CallbackBefore
-	callbackBefore map[uintptr]CallbackBefore
-
-	// mutexBefore is sync.Mutex used to sync callbackBefore
-	mutexBefore sync.Mutex
-
-	// callbackAfter is a map of:
-	//	key - sysno (uintptr)
-	//	val - CallbackAfter
-	callbackAfter map[uintptr]CallbackAfter
-
-	// mutexAfter is sync.Mutex used to sync callbackAfter
-	mutexAfter sync.Mutex
-}
-
-func (ct *CallbackTable) registerCallbackBefore(sysno uintptr, f CallbackBefore) error {
-	if f == nil {
-		return errors.New("callback func is nil")
-	}
-	ct.mutexBefore.Lock()
-	defer ct.mutexBefore.Unlock()
-
-	ct.callbackBefore[sysno] = f
-	return nil
-}
-
-func (ct *CallbackTable) registerCallbackAfter(sysno uintptr, f CallbackAfter) error {
-	if f == nil {
-		return errors.New("callback func is nil")
-	}
-	ct.mutexAfter.Lock()
-	defer ct.mutexAfter.Unlock()
-
-	ct.callbackAfter[sysno] = f
-	return nil
-}
-
-func (ct *CallbackTable) UnregisterAll() {
-	ct.mutexBefore.Lock()
-	ct.mutexAfter.Lock()
-
-	defer ct.mutexAfter.Unlock()
-	defer ct.mutexBefore.Unlock()
-
-	ct.callbackAfter = map[uintptr]CallbackAfter{}
-	ct.callbackBefore = map[uintptr]CallbackBefore{}
-}
-
-func (ct *CallbackTable) registerCallbackBeforeNoLock(sysno uintptr, f CallbackBefore) error {
-	if f == nil {
-		return errors.New("callback func is nil")
-	}
-
-	ct.callbackBefore[sysno] = f
-	return nil
-}
-
-func (ct *CallbackTable) registerCallbackAfterNoLock(sysno uintptr, f CallbackAfter) error {
-	if f == nil {
-		return errors.New("callback func is nil")
-	}
-
-	ct.callbackAfter[sysno] = f
-	return nil
-}
-
-func (ct *CallbackTable) unregisterCallbackBefore(sysno uintptr) error {
-	ct.mutexBefore.Lock()
-	defer ct.mutexBefore.Unlock()
-
-	_, ok := ct.callbackBefore[sysno]
-	if !ok {
-		return errors.New(fmt.Sprintf("before-callback with sysno %v not exist", sysno))
-	}
-
-	delete(ct.callbackBefore, sysno)
-	return nil
-}
-
-func (ct *CallbackTable) unregisterCallbackAfter(sysno uintptr) error {
-	ct.mutexAfter.Lock()
-	defer ct.mutexAfter.Unlock()
-
-	_, ok := ct.callbackAfter[sysno]
-	if !ok {
-		return errors.New(fmt.Sprintf("after-callback with sysno %v not exist", sysno))
-	}
-
-	delete(ct.callbackAfter, sysno)
-	return nil
-}
-
-func (ct *CallbackTable) getCallbackBefore(sysno uintptr) CallbackBefore {
-	ct.mutexBefore.Lock()
-
-	f, ok := ct.callbackBefore[sysno]
-	ct.mutexBefore.Unlock()
-	if ok && f != nil {
-		return f
-	} else {
-		return nil
-	}
-}
-
-func (ct *CallbackTable) getCallbackAfter(sysno uintptr) CallbackAfter {
-	ct.mutexAfter.Lock()
-
-	f, ok := ct.callbackAfter[sysno]
-	ct.mutexAfter.Unlock()
-	if ok && f != nil {
-		return f
-	} else {
-		return nil
-	}
 }
 
 // js callback staff bellow
